[3][Ahmet/Enes] Helm template todo-backend updated.
Added a StorageClass with a yaml postgre-pv to keep persistent volume
of postgre db.
Added a PersistentVolumeClaim object to control the accessing persistent
volume from postgre-db and mount the persistent volume with postgre container.
Some secret keys like postgre-user in postgre-deployment changed to names that not
contains postgre-  like password instead of postgre-password.
Wrong port exposing of api changed with correct port which is 4000